Remove legacy resource management. Fix styles.
