--- app/assets/stylesheets/manage/legacy/authors.scss
@@ -1,6 +1,0 @@
-.author-content {
-  width: 80%;
-  display: flex;
-  flex-direction: column;
-  margin: 0 auto;
-}
