--- app/assets/stylesheets/manage/legacy/citations.scss
@@ -1,11 +1,0 @@
-.content-citation-fields {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.citation-content {
-  margin: 0 auto;
-  width: 80%;
-}
