--- app/assets/stylesheets/manage/legacy/recipes.scss
@@ -1,133 +1,0 @@
-/**
-Show
- */
-
-/**
-Primary Picture
- */
-.resource-primary-img-container {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  .edit-recipe-profile-img-container {
-    padding: 2%;
-    margin: 0.5%;
-    border: 1px solid black;
-    border-radius: 30px;
-    height: 70vh;
-    width: 40%;
-    display: flex;
-    align-items: center;
-    .recipe-profile-img {
-      border-radius: 10px;
-      height: 100%;
-      width: 100%;
-    }
-  }
-
-  #primary_picture_upload_ctrl_container {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    padding: 1%;
-    border-radius: 10px;
-    width: 30%;
-    margin: 0.5% 0 0.5% 0;
-    #recipe_primary_picture {
-      margin: 0.5% auto;
-      height: 100%;
-    }
-  }
-}
-
-/**
-Gallery Pictures
- */
-
-.resource-gallery-img-container {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  border: 1px solid black;
-  margin: 1%;
-
-  .resource-gallery-preview {
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: center;
-
-    .resource-preview-container {
-      width: 45%;
-      height: 60vh;
-      margin: 1%;
-      padding: 1%;
-      border: 1px solid black;
-      border-radius: 10px;
-
-      .gallery-images-container {
-        list-style: none;
-        padding-left: 0;
-        display: flex;
-        flex-direction: row;
-        justify-content: center;
-        align-items: center;
-        flex-wrap: wrap;
-        height: 85%;
-        overflow-y: scroll;
-        .preview-resource-image {
-          margin: 1%;
-          width: 30%;
-          padding: 1%;
-          border: 1px solid black;
-          border-radius: 10px;
-          display: flex;
-          flex-direction: column;
-          justify-content: center;
-          align-items: center;
-          height: 75%;
-
-          .resource-gallery-img {
-            height: 75%;
-            width: 100%;
-            border-radius: 10px;
-          }
-        }
-      }
-    }
-  }
-
-  #gallery_pictures_ctrl {
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-  }
-}
-
-.cancel-upload-btn {
-  background-color: red;
-  border: 1px solid #861111;
-
-  &:hover {
-    background-color: #861111;
-  }
-}
-
-.delete-img-container {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  height: 20%;
-  padding: 1%;
-  .delete-img-txt {
-    padding: 0;
-    margin: 0;
-  }
-}
