--- app/assets/stylesheets/manage/legacy/shared.scss
@@ -1,179 +1,0 @@
-.directory-actions-links {
-  margin: 0 auto;
-  width: 65%;
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-end;
-  .action-links {
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-end;
-    align-items: center;
-    padding-left: 0;
-    list-style: none;
-    .directory-action-link {
-      padding: 2%;
-      text-align: center;
-      width: 20%;
-    }
-  }
-}
-
-.resource-heading {
-  text-align: center;
-}
-
-.resources-container {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.resources {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 80%;
-  .resource-list-container {
-    width: 100%;
-    .resource-list {
-      width: 100%;
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-
-      .resource-item {
-        width: 100%;
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-
-        .resource-text {
-          width: 75%;
-        }
-
-        .manage-resource-links {
-          display: flex;
-          flex-direction: row;
-          justify-content: space-evenly;
-          width: 20%;
-        }
-      }
-    }
-  }
-
-  .add-resource-container {
-    width: 100%;
-  }
-}
-
-.manage-resource-heading {
-  text-align: center;
-}
-
-.return-link {
-  padding: 2%;
-  text-align: end;
-}
-
-.action-links-navbar {
-  width: 100%;
-
-  .action-links {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    padding-left: 0;
-    list-style: none;
-    width: 100%;
-
-    .action-link {
-      display: flex;
-      flex-direction: column;
-      justify-content: center;
-      width: 15%;
-      border: 1px solid black;
-      text-align: center;
-      padding: 1%;
-    }
-  }
-}
-
-.resource-list-container.small-list-container {
-  width: 75%;
-}
-
-/**
-Form
- */
-
-.resource-form {
-  margin: 0 auto;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  .form-heading {
-    text-align: center;
-  }
-
-  .form-ctrl-container {
-    display: flex;
-    flex-direction: column;
-    width: 80%;
-
-    .form-ctrl-date {
-
-    }
-  }
-
-  .recipe-profile-img-container {
-    padding: 2%;
-    margin: 1%;
-    border: 1px solid black;
-    border-radius: 30px;
-    height: 50vh;
-    width: 40%;
-    display: flex;
-    align-self: center;
-    .recipe-profile-img {
-      border-radius: 10px;
-      height: 100%;
-      width: 100%;
-    }
-  }
-
-  .submit-container {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-  }
-}
-
-/**
-Pagination
- */
-
-.paginated-resources-container {
-  display: flex;
-  flex-direction: column;
-
-  .page-entries-info {
-    width: 40%;
-    align-self: end;
-  }
-
-  .pagination {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    span {
-      margin: 0 0.5%;
-    }
-  }
-}
