--- app/assets/stylesheets/manage/legacy/step_ingredients.scss
@@ -1,4 +1,0 @@
-.step-ingredient-content {
-  margin: 0 auto;
-  width: 80%;
-}
