--- app/assets/stylesheets/manage/legacy/steps.scss
@@ -1,13 +1,0 @@
-.step-content {
-  margin: 0 auto;
-  width: 80%;
-}
-
-.recipe-step {
-  cursor: grab;
-  border: 1px solid black;
-  padding: 1%;
-  &:hover {
-    box-shadow: 6px 6px 2px 1px #2E2F30;
-  }
-}
