--- app/controllers/manage/legacy/steps_controller.rb
@@ -1,70 +1,0 @@
-# frozen_string_literal: true
-
-class Manage::Legacy::StepsController < Manage::Legacy::ApplicationController
-  def show
-    @step = Step.managed.find(params[:id])
-    authorize! @step
-
-    render :show
-  end
-
-  def new
-    @step = Step.new(recipe: Recipe.managed.friendly.find(params[:recipe_id]))
-    authorize!
-
-    render :new
-  end
-
-  def create
-    @recipe = Recipe.managed.find(params[:recipe_id])
-    @step = Step.new({ **step_params, recipe: @recipe })
-    authorize! @step
-
-    if @step.save
-      flash.notice = 'Step created successfully.'
-      redirect_to manage_legacy_recipe_step_path(@recipe.id, @step)
-    else
-      flash.now[:error] = @step.errors.full_messages
-      render :new
-    end
-  end
-
-  def edit
-    @step = Step.find(params[:id])
-    authorize! @step
-
-    render :edit
-  end
-
-  def update
-    @step = Step.find(params[:id])
-    authorize! @step
-
-    if @step.update(step_params)
-      flash.notice = 'Step updated successfully.'
-      redirect_to manage_legacy_recipe_step_path(params[:recipe_id], @step)
-    else
-      flash.now[:error] = @step.errors.full_messages
-      render :new
-    end
-  end
-
-  def destroy
-    @step = Step.where(recipe_id: params[:recipe_id]).find(params[:id])
-    authorize! @step
-
-    Step.destroy(@step.id)
-
-    redirect_to manage_legacy_recipe_path(params[:recipe_id])
-  end
-
-  private
-
-  def step_params
-    params.require(:step).permit(
-      :id,
-      :instruction,
-      :order
-    )
-  end
-end
