+++ app/controllers/manage/stable/application_controller.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class Manage::Legacy::ApplicationController < ApplicationController
+class Manage::Stable::ApplicationController < ApplicationController
   before_action :authenticate_user!
   verify_authorized
 end
