--- app/views/application/authors/_form.html.erb
@@ -1,20 +1,0 @@
-<%= form_with model: @author, class: 'resource-form', url: action do |author_form| %>
-  <%= author_form.hidden_field :citation_id, value: author_form.object.citation.id %>
-
-  <div class="form-ctrl-container">
-    <%= author_form.label :first_name, class: "form-ctrl-lbl" %>
-    <%= author_form.text_field :first_name, class: "form-ctrl" %>
-  </div>
-  <div class="form-ctrl-container">
-    <%= author_form.label :middle_name, class: "form-ctrl-lbl" %>
-    <%= author_form.text_field :middle_name, class: "form-ctrl" %>
-  </div>
-  <div class="form-ctrl-container">
-    <%= author_form.label :last_name, class: "form-ctrl-lbl" %>
-    <%= author_form.text_field :last_name, class: "form-ctrl" %>
-  </div>
-
-  <div class="submit-container">
-    <%= author_form.submit "Save", class: "button save-btn" %>
-  </div>
-<% end %>
