--- app/views/application/citations/_form.html.erb
@@ -1,80 +1,0 @@
-<%= form_with model: @citation, class: 'resource-form', url: action do |citation_form| %>
-  <%= citation_form.hidden_field :recipe_id, value: citation_form.object.recipe.id %>
-
-  <div class="form-ctrl-container">
-    <%= citation_form.label :origin, class: "form-ctrl-lbl" %>
-    <%= citation_form.select :origin, options_for_select(Citation::ORIGINS.map{|c| [c.capitalize, c]}), class: "form-ctrl" %>
-  </div>
-  <div class="form-ctrl-container">
-    <%= citation_form.label :content_type, class: "form-ctrl-lbl" %>
-    <%= citation_form.select :content_type, options_for_select(Citation::CONTENT_TYPES.map{|o| [o.capitalize, o]}, class: "form-ctrl") %>
-  </div>
-
-  <% if http_method == :patch %>
-    <div class="form-ctrl-container">
-      <%= citation_form.label :publication_title, class: "form-ctrl-lbl" %>
-      <%= citation_form.text_field :publication_title, class: "form-ctrl" %>
-    </div>
-    <div class="form-ctrl-container">
-      <%= citation_form.label :published_at, class: "form-ctrl-lbl" %>
-      <%= citation_form.date_field :published_at, class: "form-ctrl form-ctrl-date" %>
-    </div>
-
-    <section id="book_citation_fields" class="content-citation-fields">
-      <div class="form-ctrl-container">
-        <%= citation_form.label :publication_location, class: "form-ctrl-lbl" %>
-        <%= citation_form.text_field :publication_location, class: "form-ctrl" %>
-      </div>
-      <div class="form-ctrl-container">
-        <%= citation_form.label :content_location, class: "form-ctrl-lbl" %>
-        <%= citation_form.text_field :content_location, class: "form-ctrl" %>
-      </div>
-      <div class="form-ctrl-container">
-        <%= citation_form.label :publisher, class: "form-ctrl-lbl" %>
-        <%= citation_form.text_field :publisher, class: "form-ctrl" %>
-      </div>
-      <div class="form-ctrl-container">
-        <%= citation_form.label :version, class: "form-ctrl-lbl" %>
-        <%= citation_form.text_field :version, class: "form-ctrl" %>
-      </div>
-    </section>
-
-    <section id="site_citation_fields" class="content-citation-fields">
-      <div class="form-ctrl-container">
-        <%= citation_form.label :site_link, class: "form-ctrl-lbl" %>
-        <%= citation_form.url_field :site_link, class: "form-ctrl" %>
-      </div>
-      <div class="form-ctrl-container">
-        <%= citation_form.label :site_title, class: "form-ctrl-lbl" %>
-        <%= citation_form.text_field :site_title, class: "form-ctrl" %>
-      </div>
-      <div class="form-ctrl-container">
-        <%= citation_form.label :last_accessed_at, class: "form-ctrl-lbl" %>
-        <%= citation_form.date_field :last_accessed_at, class: "form-ctrl" %>
-      </div>
-    </section>
-
-    <section class="resources">
-      <div class="resource-list-container citation-authors-container">
-        <label for="citation_authors">Authors</label>
-        <ol id="citation_authors" class="resource-list">
-          <%= citation_form.fields_for :authors do |author_form| %>
-            <li class="resource-item">
-              <p class="resource-text author-display-name"><%= "#{author_form.object.last_name}, #{author_form.object.first_name}" %></p>
-              <div class="manage-resource-links">
-                <%= link_to "Edit", edit_manage_legacy_recipe_citation_author_path(author_form.object.citation.recipe, author_form.object.citation, author_form.object), class: "button" %>
-              </div>
-            </li>
-          <% end %>
-        </ol>
-        <%= link_to "Add Author", new_manage_legacy_recipe_citation_author_path(citation_form.object.recipe, citation_form.object), class: "button" %>
-      </div>
-    </section>
-  <% else %>
-    <p>Further Citation data may only be updated on an existing Citation</p>
-  <% end %>
-
-  <div class="submit-container">
-    <%= citation_form.submit "Save", class: "button save-btn" %>
-  </div>
-<% end %>
