+++ app/views/application/manage/stable/recipes/_form.html.erb
@@ -1,8 +1,10 @@
-<nav class="form-nav">
-  <ul class="form-nav-links">
-    <li class="form-nav-link"><%= link_to 'Information', edit_manage_stable_recipe_path(@recipe, tab: 'information'), data: {"turbo-frame": :"tab_content"} %></li>
-    <li class="form-nav-link"><%= link_to 'Images', edit_manage_stable_recipe_path(@recipe, tab: 'images'), data: {"turbo-frame": :"tab_content"} %></li>
-    <li class="form-nav-link"><%= link_to 'Steps', edit_manage_stable_recipe_path(@recipe, tab: 'steps'), data: {"turbo-frame": :"tab_content"} %></li>
+<nav class="action-nav">
+  <ul class="action-nav-links">
+    <% unless @recipe.new_record? %>
+      <li class="action-nav-link"><%= link_to 'Information', edit_manage_stable_recipe_path(@recipe, tab: 'information'), data: {"turbo-frame": :"tab_content"} %></li>
+      <li class="action-nav-link"><%= link_to 'Images', edit_manage_stable_recipe_path(@recipe, tab: 'images'), data: {"turbo-frame": :"tab_content"} %></li>
+      <li class="action-nav-link"><%= link_to 'Steps', edit_manage_stable_recipe_path(@recipe, tab: 'steps'), data: {"turbo-frame": :"tab_content"} %></li>
+    <% end %>
   </ul>
 </nav>
