--- app/views/application/recipes/_form.html.erb
@@ -1,69 +1,0 @@
-<%= form_with model: @recipe, class: 'resource-form', url: action do |recipe_form| %>
-  <h3 class="form-heading" id="recipe">Recipe</h3>
-  <div class="form-ctrl-container">
-    <%= recipe_form.label :name, class: "form-ctrl-lbl" %>
-    <%= recipe_form.text_field :name, class: "form-ctrl" %>
-  </div>
-
-  <%= render partial: 'application/recipes/images/primary_picture', locals: {recipe_form: recipe_form} %>
-
-  <%= render partial: 'application/recipes/images/gallery_pictures', locals: {recipe_form: recipe_form} %>
-
-  <div class="form-ctrl-container publicly-accessible-container">
-    <%= recipe_form.label :publicly_accessible, class: "form-ctrl-lbl" %>
-    <%= recipe_form.check_box :publicly_accessible, class: "form-ctrl-text" %>
-  </div>
-
-  <div class="form-ctrl-container">
-    <%= recipe_form.label :description, class: "form-ctrl-lbl" %>
-    <%= recipe_form.text_area :description, class: "form-ctrl-text" %>
-  </div>
-
-  <section class="resources">
-    <div class="col resource-col resource-list-container" id="recipe_steps_container">
-      <label for="recipe_steps">Steps</label>
-      <ol id="recipe_steps" class="resource-list draggable">
-        <%= recipe_form.fields_for :steps do |steps_form| %>
-          <li class="resource-item recipe-step">
-            <%= steps_form.hidden_field :order, class: "recipe-step-order" %>
-            <p class="resource-text recipe-step-instruction"><%= steps_form.object.instruction %></p>
-            <div class="manage-resource-links">
-              <%= link_to "Edit", edit_manage_legacy_recipe_step_path(steps_form.object.recipe, steps_form.object), class: "button" %>
-            </div>
-          </li>
-        <% end %>
-      </ol>
-      <div class="add-resource-container">
-        <% if http_method == :patch %>
-          <%= link_to "Add Step", new_manage_legacy_recipe_step_path(@recipe), class: "button" %>
-        <% else %>
-          <p>Steps may only be added to an existing Recipe.</p>
-        <% end %>
-      </div>
-    </div>
-
-    <div class="col resource-col resource-list-container" id="recipe_citations_container">
-      <label for="recipe_citations">Citations</label>
-      <ol id="recipe_citations" class="resource-list">
-        <%= recipe_form.fields_for :citations do |citations_form| %>
-          <li class="resource-item">
-            <p class="citation-publication-title resource-text"><%= citations_form.object.publication_title %></p>
-            <div class="manage-resource-links">
-              <%= link_to "Edit", edit_manage_legacy_recipe_citation_path(citations_form.object.recipe.id, citations_form.object), class: "button" %>
-            </div>
-          </li>
-        <% end %>
-      </ol>
-      <% if http_method == :patch %>
-        <%= link_to "Add Citation", new_manage_legacy_recipe_citation_path(@recipe), class: "button" %>
-      <% else %>
-        <p>Citations may only be added to an existing Recipe.</p>
-      <% end %>
-    </div>
-  </section>
-
-  <div class="submit-container">
-    <%= recipe_form.submit "Save", class: "button save-btn" %>
-  </div>
-<% end %>
-
