+++ app/views/application/recipes/_recipe.html.erb
@@ -1,3 +1,1 @@
 <li class="directory-recipe"><%= link_to recipe.name, recipe_path(recipe) %></li>
-<!--    <span><%#= link_to 'Edit', edit_recipe_path(recipe) %></span>-->
-<!--    <span><%#= link_to "Delete", recipe_path(recipe), method: :delete, data: [confirm: "Are you sure?"] %></span>-->
