--- app/views/application/recipes/images/_gallery_pictures.html.erb
@@ -1,35 +1,0 @@
-<section class="resource-gallery-img-container">
-  <div class="resource-gallery-preview">
-    <div class="resource-preview-container">
-      <label for="existing_resource_images">Existing Images</label>
-      <ul id="existing_resource_images" class="gallery-images-container">
-        <% if recipe_form.object.gallery_pictures.attached? %>
-          <% recipe_form.object.gallery_pictures.each do |picture| %>
-            <li class="preview-resource-image existing-resource-image">
-              <%#= image_tag picture.representation(resize_to_limit: [100, 100]).processed.url %>
-              <%= image_tag picture, alt: 'Gallery picture of recipe food', class: 'resource-gallery-img' %>
-              <div class="delete-img-container">
-                <p class="delete-img-txt">Delete</p>
-                <%= check_box_tag "deleted_recipe_img_ids[]", picture.id %>
-              </div>
-            </li>
-          <% end %>
-        <% end %>
-      </ul>
-    </div>
-    <div class="resource-preview-container">
-      <label for="new_resource_images">Images to Upload</label>
-      <ul id="new_resource_images" class="gallery-images-container"></ul>
-      <template id="new_resource_image_template">
-        <li class="preview-resource-image new-resource-image">
-          <img src="" alt="<% "A picture of #{recipe_form.object.name}" %>" class="resource-gallery-img synced-multi-img" />
-        </li>
-      </template>
-    </div>
-  </div>
-
-  <div class="form-ctrl-container" id="gallery_pictures_ctrl">
-    <%= recipe_form.file_field :gallery_pictures, multiple: true, class: 'synced-multi-img' %>
-    <button type="button" class="cancel-upload-btn cancel-gallery-upload-btn">Cancel</button>
-  </div>
-</section>
