--- app/views/application/recipes/images/_primary_picture.html.erb
@@ -1,15 +1,0 @@
-<section class="resource-primary-img-container">
-  <div class="edit-recipe-profile-img-container">
-    <%= primary_picture(recipe_form.object) %>
-  </div>
-  <% if recipe_form.object.primary_picture.attached? %>
-    <div class="delete-img-container">
-      <p class="delete-img-txt">Delete</p>
-      <%= check_box_tag "deleted_recipe_img_ids[]", recipe_form.object.primary_picture.id %>
-    </div>
-  <% end %>
-  <div class="form-ctrl-container primary-picture-upload-ctrl" id="primary_picture_upload_ctrl_container">
-    <%= recipe_form.file_field :primary_picture, class: 'synced-single-img recipe-primary-picture' %>
-    <button type="button" class="cancel-upload-btn cancel-primary-picture-upload-btn">Cancel</button>
-  </div>
-</section>
