--- app/views/application/step_ingredients/_form.html.erb
@@ -1,34 +1,0 @@
-<%= form_with model: @step_ingredient, class: 'resource-form', url: action do |step_ingredient_form| %>
-  <%= step_ingredient_form.hidden_field :step_id, value: step_ingredient_form.object.step.id %>
-
-  <div class="form-ctrl-container">
-    <%= step_ingredient_form.label :amount, class: "form-ctrl-lbl" %>
-    <%= step_ingredient_form.number_field :amount, min: 0, step: "any", class: "form-ctrl" %>
-  </div>
-
-  <div class="form-ctrl-container">
-    <%= step_ingredient_form.label :unit, class: "form-ctrl-lbl" %>
-    <%= step_ingredient_form.text_field :unit, class: "form-ctrl" %>
-  </div>
-
-  <div class="form-ctrl-container">
-    <%= step_ingredient_form.label :condition, class: "form-ctrl-lbl" %>
-    <%= step_ingredient_form.text_field :condition, class: "form-ctrl" %>
-  </div>
-
-  <%= step_ingredient_form.fields_for :ingredient do |ingredient_form| %>
-    <div class="form-ctrl-container">
-      <%= ingredient_form.label :name, class: "form-ctrl-lbl" %>
-      <%= ingredient_form.text_field :name, class: "form-ctrl" %>
-    </div>
-
-    <div class="form-ctrl-container">
-      <%= ingredient_form.label :description, class: "form-ctrl-lbl" %>
-      <%= ingredient_form.text_area :description, class: "form-ctrl" %>
-    </div>
-  <% end %>
-
-  <div class="submit-container">
-    <%= step_ingredient_form.submit "Save", class: "button save-btn" %>
-  </div>
-<% end %>
