--- app/views/application/steps/_form.html.erb
@@ -1,41 +1,0 @@
-<%= form_with model: @step, class: 'resource-form', url: action do |step_form| %>
-  <%= step_form.hidden_field :recipe_id, value: step_form.object.recipe.id %>
-  <%= step_form.hidden_field :order, value: new_step_order(http_method, step_form.object.order, step_form.object.recipe) %>
-  <div class="form-ctrl-container">
-    <label for="step_order" class="form-ctrl-label">Order: </label>
-    <p id="step_order" class="form-ctrl"><%= step_form.object.order %></p>
-    <p class="order-info">Steps may only be reordered from Recipe page.
-      Click <span><%= link_to "here", edit_manage_legacy_recipe_path(step_form.object.recipe) %></span> to reorder recipe steps.
-    </p>
-
-    <div class="form-ctrl-container">
-      <%= step_form.label :instruction, class: "form-ctrl-lbl" %>
-      <%= step_form.text_area :instruction, class: "form-ctrl-text" %>
-    </div>
-  </div>
-
-<section class="resources">
-  <div class="resource-list-container small-list-container">
-    <label for="recipe_step_ingredients">Step Ingredients</label>
-    <ul id="recipe_step_ingredients" class="resource-list">
-      <%= step_form.fields_for :step_ingredients do |step_ingredients_form| %>
-        <li class="resource-item">
-          <p class="step-ingredient-name resoure-text"><%= step_ingredients_form.object.ingredient.name %></p>
-          <div class="manage-resource-links">
-            <%= link_to "Edit", edit_manage_legacy_recipe_step_step_ingredient_path(step_ingredients_form.object.step.recipe.id, step_ingredients_form.object.step.id, step_ingredients_form.object), class: "button" %>
-          </div>
-        </li>
-      <% end %>
-    </ul>
-    <% if http_method == :patch %>
-      <%= link_to "Add Ingredient", new_manage_legacy_recipe_step_step_ingredient_path(step_form.object.recipe.id, step_form.object.id), class: "button" %>
-    <% else %>
-      <p>Ingredients may only be added to an existing Step.</p>
-    <% end %>
-  </div>
-</section>
-
-  <div class="submit-container">
-    <%= step_form.submit "Save", class: "button save-btn" %>
-  </div>
-<% end %>
