--- app/views/manage/legacy/authors/show.html.erb
@@ -1,24 +1,0 @@
-<% content_for :head_js do %>
-  <%= javascript_pack_tag 'manage', 'data-turbolinks-track': 'reload' %>
-<% end %>
-
-<h1 class="resource-heading"><%= "#{@author.last_name}, #{@author.first_name}" %></h1>
-
-<section class="author-content">
-  <label for="first_name">First name: </label>
-  <p id="first_name"><%= @author.first_name %></p>
-
-  <label for="middle_name">Middle name: </label>
-  <p id="first_name"><%= @author.middle_name %></p>
-
-  <label for="last_name">Last name: </label>
-  <p id="first_name"><%= @author.last_name %></p>
-</section>
-
-<nav class="action-links-navbar">
-  <ul class="action-links">
-    <li class="action-link"><span><%= link_to "Edit", edit_manage_legacy_recipe_citation_author_path(@author.citation.recipe, @author.citation, @author) %></span></li>
-    <li class="action-link"><span><%= link_to "Delete", manage_legacy_recipe_citation_path(@author.citation.recipe, @author.citation, @author), {action: :destroy, method: :delete, recipe_id: @author.citation.recipe.id, citation_id: @author.citation.id, id: @author.id, data: {confirm: "Are you sure?"}} %></span></li>
-    <li class="action-link"><span><%= link_to "Back", manage_legacy_recipe_citation_path(@author.citation.recipe, @author.citation) %></span></li>
-  </ul>
-</nav>
