--- app/views/manage/legacy/citations/citations/_book.html.erb
@@ -1,24 +1,0 @@
-<h1 class="citation-heading book-heading resource-heading">Citation: <%= @citation.publication_title %></h1>
-
-<section class="citation-content">
-  <p id="origin"><%= @citation.origin %></p>
-  <p id="content_type"><%= @citation.content_type %></p>
-
-  <label for="publication_title">Title: </label>
-  <p id="publication_title"><%= @citation.publication_title %></p>
-
-  <label for="publisher">Published by: </label>
-  <p id="publisher"><%= @citation.publisher %></p>
-
-  <label for="publication_location">Published in: </label>
-  <p id="publication_location"><%= @citation.publication_location %></p>
-
-  <label for="published_at">Published at: </label>
-  <p id="published_at"><%= @citation.published_at %></p>
-
-  <label for="content_location">Location in book/content: </label>
-  <p id="content_location"><%= @citation.content_location %></p>
-
-  <label for="version">Version: </label>
-  <p id="version"><%= @citation.version %></p>
-</section>
