--- app/views/manage/legacy/citations/citations/_site.html.erb
@@ -1,27 +1,0 @@
-<h1 class="citation-heading site-heading resource-heading">Citation: <%= @citation.site_title %></h1>
-
-<section class="citation-content">
-  <label for="origin">Recipe origin: </label>
-  <p id="origin"><%= @citation.origin %></p>
-
-  <label for="content_type">Type of recipe: </label>
-  <p id="content_type"><%= @citation.content_type %></p>
-
-  <label for="publication_title">Title: </label>
-  <p id="publication_title"><%= @citation.publication_title %></p>
-
-  <label for="site_title">Web Site Title: </label>
-  <p id="site_title"><%= @citation.site_title %></p>
-
-  <label for="last_accessed_at">Date last accessed at: </label>
-  <p id="last_accessed_at"><%= @citation.last_accessed_at %></p>
-
-  <label for="site_link">Link to site: </label>
-  <p id="site_link"><%= @citation.site_link %></p>
-
-  <label for="publisher">Published by: </label>
-  <p id="publisher"><%= @citation.publisher %></p>
-
-  <label for="published_at">Published at: </label>
-  <p id="published_at"><%= @citation.published_at %></p>
-</section>
