--- app/views/manage/legacy/citations/edit.html.erb
@@ -1,9 +1,0 @@
-<% content_for :head_js do %>
-  <%= javascript_pack_tag 'manage', 'data-turbolinks-track': 'reload' %>
-<% end %>
-
-<h1 class="manage-resource-heading">Edit Citation</h1>
-
-<%= render partial: "application/citations/form", locals: {http_method: :patch, action: manage_legacy_recipe_citation_path(@citation.recipe.id, @citation)} %>
-
-<div class="return-link"><%= link_to "Back", manage_legacy_recipe_citation_path(@citation.recipe.id, @citation) %></div>
