--- app/views/manage/legacy/recipes/index.html.erb
@@ -1,27 +1,0 @@
-<% content_for :head_js do %>
-  <%= javascript_pack_tag 'manage', 'data-turbolinks-track': 'reload' %>
-<% end %>
-
-<h1 class="recipes-heading">Listing Recipes</h1>
-
-<div class="paginated-resources-container">
-  <p class="page-entries-info">
-    <%= page_entries_info @recipes %>
-  </p>
-
-  <ul id="directory_recipes">
-    <% @recipes.each do |recipe| %>
-      <li class="directory-recipe">
-        <span><%= link_to recipe.name, manage_legacy_recipe_path(recipe) %></span>
-        <div class="recipe-links">
-          <span><%= link_to "Edit", edit_manage_legacy_recipe_path(recipe) %></span>
-          <span><%= link_to "Delete", manage_legacy_recipe_path(recipe), {action: :destroy, method: :delete, id: recipe.id, data: {confirm: "Are you sure?"}} %></span>
-        </div>
-      </li>
-    <% end %>
-  </ul>
-
-  <%= paginate @recipes %>
-</div>
-
-<div class="new-recipe-link"><%= link_to "New Recipe", new_manage_legacy_recipe_path %></div>
