--- app/views/manage/legacy/recipes/new.html.erb
@@ -1,9 +1,0 @@
-<% content_for :head_js do %>
-  <%= javascript_pack_tag 'manage', 'data-turbolinks-track': 'reload' %>
-<% end %>
-
-<h1 class="manage-resource-heading">New Recipe</h1>
-
-<%= render partial: "application/recipes/form", locals: {http_method: :post, action: manage_legacy_recipes_path} %>
-
-<div class="return-link"><%= link_to "Back", manage_legacy_recipes_path %></div>
