--- app/views/manage/legacy/recipes/show.html.erb
@@ -1,54 +1,0 @@
-<% content_for :head_js do %>
-  <%= javascript_pack_tag 'manage', 'data-turbolinks-track': 'reload' %>
-<% end %>
-
-<h1 class="resource-heading">Manage: <%= @recipe.name %></h1>
-
-<section class="recipe-bio">
-  <div class="recipe-profile-img-container">
-    <%= primary_picture(@recipe) %>
-  </div>
-  <p aria-label="Recipe description" id="recipe_description"><%= @recipe.description %></p>
-  <p aria-label="Recipe publicly accessible" id="publicly_accessible"><%= publicly_accessible_display(@recipe.publicly_accessible) %></p>
-</section>
-
-<div class="resources-container">
-  <section class="resources">
-    <div class="resource-list-container">
-      <label for="recipe_steps">Steps</label>
-      <ul id="recipe_steps" class="resource-list">
-        <% @recipe.steps.each do |step| %>
-          <li class="recipe-step resource-item">
-            <p class="recipe-step-instruction resource-text"><%= step.instruction %></p>
-            <div class="recipe-step-links manage-resource-links">
-              <span><%= link_to "Show", manage_legacy_recipe_step_path(step.recipe.id, step) %></span>
-              <span><%= link_to "Edit", edit_manage_legacy_recipe_step_path(step.recipe.id, step) %></span>
-            </div>
-          </li>
-        <% end %>
-      </ul>
-    </div>
-    <div class="resource-list-container">
-      <label for="recipe_citations">Citations</label>
-      <ul id="recipe_citations" class="resource-list">
-        <% @recipe.citations.each do |citation| %>
-          <li class="recipe-citation resource-item">
-            <p class="recipe-citation-publication-title resource-text"><%= citation.publication_title %></p>
-            <div class="recipe-citation-links manage-resource-links">
-              <div><%= link_to "Show", manage_legacy_recipe_citation_path(citation.recipe.id, citation) %></div>
-              <div><%= link_to "Edit", edit_manage_legacy_recipe_citation_path(citation.recipe.id, citation) %></div>
-            </div>
-          </li>
-        <% end %>
-      </ul>
-    </div>
-  </section>
-</div>
-
-<nav class="action-links-navbar">
-  <ul class="action-links">
-    <li class="action-link"><%= link_to "Edit", edit_manage_legacy_recipe_path(@recipe) %></li>
-    <li class="action-link"><%= link_to "Delete", manage_legacy_recipe_path(@recipe), {action: :destroy, method: :delete, id: @recipe.id, data: {confirm: "Are you sure?"}} %></li>
-    <li class="action-link"><%= link_to "Back", manage_legacy_recipes_path %></li>
-  </ul>
-</nav>
