--- app/views/manage/legacy/steps/edit.html.erb
@@ -1,9 +1,0 @@
-<% content_for :head_js do %>
-  <%= javascript_pack_tag 'manage', 'data-turbolinks-track': 'reload' %>
-<% end %>
-
-<h1 class="manage-resource-heading">Edit Step</h1>
-
-<%= render partial: "application/steps/form", locals: {http_method: :patch, action: manage_legacy_recipe_step_path(@step.recipe.id, @step)} %>
-
-<div class="return-link"><%= link_to "Back", manage_legacy_recipe_step_path(@step.recipe.id, @step) %></div>
