--- app/views/manage/legacy/steps/show.html.erb
@@ -1,38 +1,0 @@
-<% content_for :head_js do %>
-  <%= javascript_pack_tag 'manage', 'data-turbolinks-track': 'reload' %>
-<% end %>
-
-<h1 class="resource-heading">Step <%= @step.order %></h1>
-
-<section class="step-content">
-  <label for="step_instruction">Instruction: </label>
-  <p id="step_instruction"><%= @step.instruction %></p>
-</section>
-
-<div class="resources-container">
-  <section class="resources">
-    <div class="resource-list-container small-list-container">
-      <label for="recipe_step_ingredients">Ingredients</label>
-      <ul id="recipe_step_ingredients" class="resource-list">
-        <% @step.step_ingredients.each do |step_ingredient| %>
-          <li class="step-ingredient resource-item">
-            <p class="step-ingredient-name resource-text"><%= step_ingredient.ingredient.name %></p>
-            <div class="step-ingredient-links manage-resource-links">
-              <span><%= link_to "Show", manage_legacy_recipe_step_step_ingredient_path(step_ingredient.step.recipe.id, step_ingredient.step.id, step_ingredient) %></span>
-              <span><%= link_to "Edit", edit_manage_legacy_recipe_step_step_ingredient_path(step_ingredient.step.recipe.id, step_ingredient.step.id, step_ingredient) %></span>
-              <span><%= link_to "Delete", manage_legacy_recipe_step_step_ingredient_path(step_ingredient.step.recipe.id, step_ingredient.step.id, step_ingredient), {action: :destroy, method: :delete, recipe_id: step_ingredient.step.recipe.id, step_id: step_ingredient.id, id: step_ingredient.id, data: {confirm: "Are you sure?"}} %></span>
-            </div>
-          </li>
-        <% end %>
-      </ul>
-    </div>
-  </section>
-</div>
-
-<nav class="action-links-navbar">
-  <ul class="action-links">
-    <li class="action-link"><span><%= link_to "Edit", edit_manage_legacy_recipe_step_path(@step.recipe_id, @step) %></span></li>
-    <li class="action-link"><span><%= link_to "Delete", manage_legacy_recipe_step_path(@step.recipe_id, @step), {action: :destroy, method: :delete, recipe_id: @step.recipe.id, id: @step.id, data: {confirm: "Are you sure?"}} %></span></li>
-    <li class="action-link"><span><%= link_to "Back", manage_legacy_recipe_path(@step.recipe_id) %></span></li>
-  </ul>
-</nav>
