+++ app/views/manage/stable/recipes/new.html.erb
@@ -2,4 +2,4 @@
 
 <%= render partial: 'application/manage/stable/recipes/form', locals: {action: manage_stable_recipes_path} %>
 
-<p>Note: Only basic information may be added for new recipes</p>
+<p class="new-resource-note">Note: Only basic information may be added for new recipes</p>
