+++ app/views/manage/stable/recipes/show.html.erb
@@ -20,10 +20,10 @@
   </section>
 </div>
 
-<nav>
-  <ul class="action-links">
-    <li class="action-link"><%= link_to "Edit", edit_manage_stable_recipe_path(@recipe) %></li>
-    <li class="action-link"><%= link_to "Delete", manage_stable_recipe_path(@recipe), {action: :destroy, method: :delete, id: @recipe.id, data: {confirm: "Are you sure?"}} %></li>
-    <li class="action-link"><%= link_to "Back", manage_stable_recipes_path %></li>
+<nav class="action-nav">
+  <ul class="action-nav-links">
+    <li class="action-nav-link"><%= link_to "Edit", edit_manage_stable_recipe_path(@recipe) %></li>
+    <li class="action-nav-link"><%= link_to "Delete", manage_stable_recipe_path(@recipe), {action: :destroy, method: :delete, id: @recipe.id, data: {confirm: "Are you sure?"}} %></li>
+    <li class="action-nav-link"><%= link_to "Back", manage_stable_recipes_path %></li>
   </ul>
 </nav>
