+++ app/views/recipes/index.html.erb
@@ -15,7 +15,7 @@
 <div class="directory-actions-links">
   <ul class="action-links">
     <li class="directory-action-link">
-      <%= link_to "New Recipe", new_manage_legacy_recipe_path %>
+      <%= link_to "New Recipe", new_manage_stable_recipe_path %>
     </li>
   </ul>
 </div>
