+++ app/views/recipes/show.html.erb
@@ -2,7 +2,7 @@
 
 <section class="recipe-content">
   <% if user_signed_in? && allowed_to?(:edit?, @recipe) %>
-    <%= link_to 'Edit', edit_manage_legacy_recipe_path(@recipe) %>
+    <%= link_to 'Edit', edit_manage_stable_recipe_path(@recipe) %>
   <% end %>
 
   <div class="recipe-profile-img-container">
