+++ config/routes.rb
@@ -3,18 +3,6 @@ Rails.application.routes.draw do
   root to: 'landings#home'
 
   namespace :manage do
-    namespace :legacy do
-      resources :recipes do
-        resources :steps, except: :index do
-          resources :step_ingredients, except: :index
-        end
-
-        resources :citations, except: :index do
-          resources :authors, except: :index
-        end
-      end
-    end
-
     namespace :stable do
       resources :recipes do
         resources :steps
